Remove layers

+++ lumina_next_t2i_mini/train_vae_swap_layers.py
@@ -192,7 +192,7 @@ def main(args):
 
 
     # Model compile and checkpoint
-    model.half()
+    # model.half()
 
     # Optimizer
     logger.info(f"Creating optimizer")
